requirements.txt.txt
streamlit==1.32.2
pandas==2.2.2
requests==2.31.0
googlesearch-python==1.1.0
